RobotMotionManager - option to disable movement (e.g. with open settings)
