+++ Assets/Scripts/InputDevices/VRControllers/VRControllerInputSystem.cs
@@ -70,6 +70,8 @@ namespace InputDevices.VRControllers
         /// <param name="actionToInvoke"></param>
         private void HandleButtonPress(bool isPressed, ref bool previouslyPressed, Action actionToInvoke)
         {
+            // TODO: fix pressing buttons, this doesn't work properly
+            Debug.Log($"buttons: {isPressed}, {previouslyPressed}");
             if (isPressed)
             {
                 if (!previouslyPressed)
